requirements.txt
jittor==1.3.2.7
matplotlib==3.5.1
numpy==1.19.5
opencv_python==4.5.5.64
Pillow==9.2.0
scipy==1.7.3
tqdm==4.64.0
